Tomb exit and code compacting
Tidying up a lot of scripts, to make them shorter, more compact. The program is meeting the memory limit for the _TEXT segment, and because of the integration with the assembly language, the solution is not trivial. So lots of tidying.

Also, started to add code for entombing the 3 artifacts (skull, ring, wand) and the resulting effects.