Klimb and Descend plus CollapseCrypt and OrbTest
Polished up a lot of stuff. Fixed how the final closing of Crypts and Oracles is handled. Made a new function, Cryptcollapse, that handles closing crypts. Crypts will now close no matter the direction in which the player moves after leaving them, and even if he 'blinks' away. Oracles will stay open whether an orb or an artifact is put in them. However, when Orbs are put in the chance of Daemons appearing in random encounters increases. A new 'OrbCheck' function was added save some space when checking for all use of all three orbs.
'Views' were added for all towns when looking through the telescope in the Lycaeum, but not for any second floors.
Handling of second floors and basements was improved, and now allows for basements or 2nd floors to be retroactively added in, provided the naming is correct. The code automatically pickup ladders and links the floors accurately. The handling of Talk files was also integrated with this, and works smoothly after the addition of another variable for basements' talk files
2nd floors and Basements were added for castles and several towns, as well as talk files for multiple towns. That still needs to be finished.